feat(telemetry): move metrics routers to domain
