+++ apps/backend/app/api/__init__.py
@@ -1,0 +1,5 @@
+from __future__ import annotations
+
+from app.domains.telemetry.api import rum_metrics_router as rum_metrics
+
+__all__ = ["rum_metrics"]
